separate login & view
