+++ src/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
-import App from './components/App'
+import AppScreenContainer from './screens/AppScreen/AppScreenContainer'
 
 ReactDOM.render (
-    <App />,
+    <AppScreenContainer />,
     document.getElementById("root")
 );
