+++ tsconfig.json
@@ -16,7 +16,7 @@
       "baseUrl": "./",
       "paths": {
         "components/*": [
-          "src/components/*"
+          "src/Screens/*"
         ],
       },
       "noEmit": true,
